cloud_function_example/requirements.txt
# Function dependencies, for example:
# package>=version
flask
numpy
google-cloud-storage